+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## DEV
 
+ * Show root pages in the custom navigation module (see contao/core-bundle#1641).
  * Add the `removeField()` method to the palette manipulator (see contao/core-bundle#1668).
  * Allow to select a news reader in the news list (see contao/news-bundle#39).
  * Show the current page title and URL in the preview bar (see contao/core-bundle#1640).